# SOLR-16097 Batch add docs when starting tests