--- a/solr/core/src/test/org/apache/solr/handler/component/TestDistributedStatsComponentCardinality.java
+++ b/solr/core/src/test/org/apache/solr/handler/component/TestDistributedStatsComponentCardinality.java
@@ -22,14 +22,18 @@ import java.lang.invoke.MethodHandles;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.stream.IntStream;
 import org.apache.lucene.util.LuceneTestCase.Slow;
 import org.apache.lucene.util.TestUtil;
 import org.apache.solr.BaseDistributedSearchTestCase;
 import org.apache.solr.SolrTestCaseJ4.SuppressSSL;
 import org.apache.solr.client.solrj.response.FieldStatsInfo;
 import org.apache.solr.client.solrj.response.QueryResponse;
+import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.util.LogLevel;
@@ -77,27 +81,32 @@ public class TestDistributedStatsComponentCardinality extends BaseDistributedSea
     log.info("Building an index of {} docs", NUM_DOCS);
 
     // we want a big spread in the long values we use, decrement by BIG_PRIME as we index
-    long longValue = MAX_LONG;
+    final AtomicLong longValue = new AtomicLong(MAX_LONG);
 
-    for (int i = 1; i <= NUM_DOCS; i++) {
-      // with these values, we know that every doc indexed has a unique value in all of the
-      // fields we will compute cardinality against.
-      // which means the number of docs matching a query is the true cardinality for each field
+    Iterator<SolrInputDocument> docs =
+        IntStream.range(1, NUM_DOCS + 1)
+            .mapToObj(
+                i -> {
+                  long currentLong = longValue.getAndAccumulate(BIG_PRIME, (x, y) -> x - y);
 
-      final String strValue = "s" + longValue;
-      indexDoc(
-          sdoc(
-              "id", "" + i,
-              "int_i", "" + i,
-              "int_i_prehashed_l", "" + HASHER.hashInt(i).asLong(),
-              "long_l", "" + longValue,
-              "long_l_prehashed_l", "" + HASHER.hashLong(longValue).asLong(),
-              "string_s", strValue,
-              "string_s_prehashed_l",
-                  "" + HASHER.hashString(strValue, StandardCharsets.UTF_8).asLong()));
+                  final String strValue = "s" + currentLong;
 
-      longValue -= BIG_PRIME;
-    }
+                  // with these values, we know that every doc indexed has a unique value in all of
+                  // the fields we will compute cardinality against. which means the number of docs
+                  // matching a query is the true cardinality for each field
+                  return sdoc(
+                      "id", Integer.toString(i),
+                      "int_i", Integer.toString(i),
+                      "int_i_prehashed_l", Long.toString(HASHER.hashInt(i).asLong()),
+                      "long_l", Long.toString(currentLong),
+                      "long_l_prehashed_l", Long.toString(HASHER.hashLong(currentLong).asLong()),
+                      "string_s", strValue,
+                      "string_s_prehashed_l",
+                          Long.toString(
+                              HASHER.hashString(strValue, StandardCharsets.UTF_8).asLong()));
+                })
+            .iterator();
+    indexDocs(docs);
 
     commit();
   }