https://issues.apache.org/jira/browse/SOLR-16097

TestDistributedStatsComponentCardinality gets 40% faster (29 -> 17s)
DistribCursorPagingTest gets 20% faster (24 -> 20 s)